# src/openApi.yml
swagger: "2.0"
info:
  description: "This is a Book Store server"
  version: "1.0.0"
  title: "Book Store"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "uvak777@gamil.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "bookstore.swagger.io"
basePath: "/v2"
...
schemes:
  - "https"
  - "http"
paths:
...
  /book:
    post:
      tags:
        - "book"
      summary: "Add a new book to the store"
      description: ""
      operationId: "addBook"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Book object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/book"
      responses:
        "200":
          description: "successfully saved book"
          schema:
            $ref: "#/definitions/book"
        "405":
          description: "Invalid input"
  /book/{bookID}:
    get:
      tags:
        - "book"
      summary: "Find book by ID"
      description: "Returns a single book"
      operationId: "getBookById"
      produces:
        - "application/json"
      parameters:
        - name: "bookID"
          in: "path"
          description: "ID of book to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/book"
        "404":
          description: "Book not found"
    put:
      tags:
        - "book"
      summary: "Updates a book in the store with form data"
      description: ""
      operationId: "updateBookWithForm"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "bookID"
          in: "path"
          description: "ID of book that needs to be updated"
          required: true
          type: "integer"
          format: "int64"
        - name: "bookName"
          in: "formData"
          description: "Updated name of the book"
          required: false
          type: "string"
        - name: "publisher"
          in: "formData"
          description: "Updated publisher of the book"
          required: false
          type: "string"
        - name: "numberOfPages"
          in: "formData"
          description: "Updated number of pages of the book"
          required: true
          type: "integer"
        - name: "year"
          in: "formData"
          description: "Updated year of the book"
          required: true
          type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/book"
        "405":
          description: "Invalid input"
    delete:
      tags:
        - "book"
      summary: "Deletes a book"
      description: ""
      operationId: "deleteBook"
      produces:
        - "application/json"
      parameters:
        - name: "bookID"
          in: "path"
          description: "Book id to delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "404":
          description: "Book not found"
...
definitions:
  author:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      authorSurname:
        type: "string"
        format: "byte"
      authorName:
        type: "string"
        format: "byte"
      authorPatronymic:
        type: "string"
        format: "byte"
      birthDate:
        type: "string"
        format: "date"
      writtenBooks:
        type: "array"
        items:
          $ref: "#/definitions/book"
  publisher:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      publisherName:
        type: "string"
        format: "byte"
      publishedBooks:
        type: "array"
        items:
          $ref: "#/definitions/book"
  book:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      price:
        type: "number"
        format: "double"
      authorsOfBook:
        type: "array"
        items:
          $ref: "#/definitions/author"
      publisher:
        $ref: "#/definitions/publisher"
      bookName:
        type: "string"
        format: "byte"
      numberOfPages:
        type: "integer"
        format: "int64"
      year:
        type: "integer"
        format: "int64"